+++ src/main/java/ru/aston/mcs/service/HousekeeperService.java
@@ -1,15 +1,12 @@
 package ru.aston.mcs.service;
 
 import ru.aston.mcs.dto.HousekeeperDTO;
-import ru.aston.mcs.entity.Housekeeper;
-
-import java.util.List;
 
 public interface HousekeeperService {
 
     void addAndSaveHousekeeper(HousekeeperDTO housekeeperDTO);
 
-    void deleteHousekeeper(int housekeeperId);
+    void deleteHousekeeper(Long housekeeperId);
 
-    HousekeeperDTO getHousekeeper(int housekeeperId);
+    HousekeeperDTO getHousekeeper(Long housekeeperId);
 }
